requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.37.1
django-appconf==1.0.2
django-avatar==4.1.0
django-bootstrap-form==3.4
django-bootstrap4==0.0.7
django-braces==1.13.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-hstore==1.4.2
django-jet==1.0.8
django-node-assets==0.9.2
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
django-template-check==0.3.1
django-tables2==2.0.2
djangorestframework==3.8.2
djangorestframework-gis==0.13
djangorestframework-hstore==1.3
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauth2-provider==0.0
oauthlib==2.1.0
openapi-codec==1.3.2
passlib==1.7.1
Pillow==5.3.0
pipreqs==0.4.9
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
yarg==0.1.9
